added files for soon-to-be modules and selection to markovmodel module
