+++ gpvolve/__init__.py
@@ -1,5 +1,7 @@
 from .__version__ import __version__
 
-from .markovmodel import EvoMarkovStateModel
+from .markovmodel import EvoMSM
+from .old_markovmodel import EvoMarkovStateModel
 from .visualization import *
 from .utils import *
+from .fitness import *
